Period label component
