+++ src/components/map/MapProvider.js
@@ -1,7 +1,7 @@
 import { Component } from 'react';
 import PropTypes from 'prop-types';
-import D2map from '@dhis2/gis-api';
-// import D2map from '@dhis2/maps-gl';
+// import D2map from '@dhis2/gis-api';
+import D2map from '@dhis2/maps-gl';
 
 // Makes the Leaflet map instance available in all child components
 class MapProvider extends Component {
